gui: Tidy up source files
