+++ src/emulator/gui/CMakeLists.txt
@@ -1,25 +1,25 @@
 add_library(
 gui
 STATIC
-include/gui/imgui_impl_sdl_gl3.h
-include/gui/imgui_impl.h
 include/gui/functions.h
+include/gui/imgui_impl_sdl_gl3.h
 include/gui/state.h
+src/common_dialog.cpp
+src/condvars_dialog.cpp
+src/eventflags_dialog.cpp
+src/game_selector.cpp
+src/gui.cpp
 src/imgui_impl_sdl_gl3.cpp
-src/imgui_impl.cpp
-src/ui_common_dialog.cpp
-src/ui_condvars_dialog.cpp
-src/ui_eventflags_dialog.cpp
-src/ui_main.cpp
-src/ui_main_menubar.cpp
-src/ui_mutexes_dialog.cpp
-src/ui_private.h
-src/ui_semaphores_dialog.cpp
-src/ui_threads_dialog.cpp
-src/ui_controls_dialog.cpp
-src/ui_allocations_dialog.cpp
-src/ui_disassembly_dialog.cpp
-src/ui_about_dialog.cpp
+src/main_menubar.cpp
+src/mutexes_dialog.cpp
+src/private.h
+src/reinstall.cpp
+src/semaphores_dialog.cpp
+src/threads_dialog.cpp
+src/controls_dialog.cpp
+src/allocations_dialog.cpp
+src/disassembly_dialog.cpp
+src/about_dialog.cpp
 )
 
 target_include_directories(gui PUBLIC include)
