+++ src/emulator/gui/src/gui.cpp
@@ -15,7 +15,9 @@
 // with this program; if not, write to the Free Software Foundation, Inc.,
 // 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
-#include <gui/imgui_impl.h>
+#include <gui/functions.h>
+
+#include "private.h"
 
 #include <gui/imgui_impl_sdl_gl3.h>
 
@@ -26,7 +28,6 @@
 #include <util/string_utils.h>
 
 #include <SDL_video.h>
-#include <imgui.h>
 
 #include <fstream>
 #include <string>
@@ -141,6 +142,49 @@ void draw_end(SDL_Window *window) {
     ImGui::Render();
     ImGui_ImplSdlGL3_RenderDrawData(ImGui::GetDrawData());
     SDL_GL_SwapWindow(window);
+}
+
+void draw_ui(HostState &host) {
+    draw_main_menu_bar(host);
+    
+    ImGui::PushFont(host.gui.monospaced_font);
+    if (host.gui.threads_dialog) {
+        draw_threads_dialog(host);
+    }
+    if (host.gui.thread_details_dialog) {
+        draw_thread_details_dialog(host);
+    }
+    if (host.gui.semaphores_dialog) {
+        draw_semaphores_dialog(host);
+    }
+    if (host.gui.mutexes_dialog) {
+        draw_mutexes_dialog(host);
+    }
+    if (host.gui.lwmutexes_dialog) {
+        draw_lw_mutexes_dialog(host);
+    }
+    if (host.gui.condvars_dialog) {
+        draw_condvars_dialog(host);
+    }
+    if (host.gui.lwcondvars_dialog) {
+        draw_lw_condvars_dialog(host);
+    }
+    if (host.gui.eventflags_dialog) {
+        draw_event_flags_dialog(host);
+    }
+    if (host.gui.controls_dialog) {
+        draw_controls_dialog(host);
+    }
+    if (host.gui.allocations_dialog) {
+        draw_allocations_dialog(host);
+    }
+    if (host.gui.disassembly_dialog) {
+        draw_disassembly_dialog(host);
+    }
+    if (host.gui.about_dialog) {
+        draw_about_dialog(host);
+    }
+    ImGui::PopFont();
 }
 
 } // namespace gui
