+++ src/emulator/gui/src/private.h
